requirements.txt
# Core dependencies
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pydantic>=2.0.0

# Build tools
pyinstaller>=6.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Windows-specific (will be ignored on other platforms)
pywin32>=306; sys_platform == "win32"

# Optional: for advanced features
click>=8.0.0
colorama>=0.4.0